vision/requirements.txt
opencv-python-headless
requests
pyyaml
python-dotenv
suntime
numpy
google-genai
